Split and fix routes
